[direct] move shiftedWhile back to Kyo.*
