+++ kyo-kernel/shared/src/main/scala/kyo/Kyo.scala
@@ -756,4 +756,37 @@ object Kyo:
         source match
             case seq: IndexedSeq[A] => seq
             case other              => Chunk.from(other)
+
+    // for kyo-direct
+    private[kyo] def shiftedWhile[A, S, B, C](source: IterableOnce[A])(
+        prolog: B,
+        f: Safepoint ?=> A => Boolean < S,
+        acc: (B, Boolean, A) => B,
+        epilog: B => C
+    )(using Frame, Safepoint): C < S =
+        source.knownSize match
+            case 0 => epilog(prolog)
+            case _ =>
+                source match
+                    case linearSeq: LinearSeq[A] => Loop(linearSeq, prolog): (seq, b) =>
+                            if seq.isEmpty then Loop.done(epilog(b))
+                            else
+                                val curr = seq.head
+                                f(curr).map:
+                                    case true  => Loop.continue(seq.tail, acc(b, true, curr))
+                                    case false => Loop.done(epilog(acc(b, false, curr)))
+
+                    case other =>
+                        val indexedSeq = toIndexed(other)
+                        val size       = indexedSeq.length
+                        Loop.indexed(prolog): (idx, b) =>
+                            if idx == size then Loop.done(epilog(b))
+                            else
+                                val curr = indexedSeq(idx)
+                                f(curr).map:
+                                    case true  => Loop.continue(acc(b, true, curr))
+                                    case false => Loop.done(epilog(acc(b, false, curr)))
+
+        end match
+    end shiftedWhile
 end Kyo
